feat: decouple reactive calcs
