+++ modules/laguerre.py
@@ -1,23 +1,94 @@
-import pathlib
-from dataclasses import asdict
-from typing import Generator
-
 import numpy as np
 import pandas as pd
-from shiny import Inputs, Outputs, Session, module, reactive, render, ui
-from shiny.types import FileInfo, ImgData
+from shiny import Inputs, ui
 from shiny_validate import InputValidator
-from synthetmic import LaguerreDiagramGenerator
-from synthetmic.data.utils import SynthetMicData, sample_random_seeds
 
-from modules import common
 from shared import comps, utils, views
 from shared.consts import (
-    Dimension,
+    APP_NAME,
     Distribution,
+    ExampleDataName,
     Phase,
-    SeedInitializer,
+    PropertyExtension,
 )
+
+
+def help(data_selection_id: str, data_extension_id: str, data_card_id: str) -> None:
+    ui.modal_show(
+        ui.modal(
+            ui.markdown(f"### {APP_NAME}"),
+            ui.hr(),
+            ui.markdown(
+                """
+                #### Usage
+
+                Using the app is extremely easy. It can be done in **4 steps**:
+                1. Specify both the space and box dimension. You can optionally
+                choose whether the underlying domain should be periodic in all directions.
+                1. Specify the number of grains and how target volumes will be generated.
+                We support single and dual phase volumes specification. You
+                can also upload your custom target volumes instead.
+                1. Click on **Generate miscrostructure** button to generate synthetic miscrostructure.
+                1. Click on any of the **download buttons** to either download
+                the generated diagram (in different formats!) or the diagram properties
+                (like centroids, vertices, etc; also in differnt formats!).
+
+                That is it!
+
+                Enjoy generating microstructures!
+                """
+            ),
+            ui.hr(),
+            ui.markdown("#### Starting point"),
+            ui.markdown(
+                """
+                Don't know where to start yet?
+
+                Don't worry, we've got you covered!
+
+                Select and download one of our example data below. You can then upload them
+                in the main app to generate the corresponding microstructure. When you click the 
+                download button, the seeds, volumes and domain of the underlying microstructure will be
+                downloaded in the selected format, which are zipped. Unzip to access the files.
+
+                Ensure you enter the correct dimension in the 'Box dimension' input. You can read this from 
+                the dimension.txt or dimension.csv file.
+                """
+            ),
+            ui.row(
+                ui.column(
+                    4,
+                    ui.card(
+                        ui.card_header("Options center"),
+                        comps.create_selection(
+                            id=data_selection_id,
+                            label="Choose an example data to download",
+                            choices=[e for e in ExampleDataName],
+                            selected=ExampleDataName.BASIC,
+                            width="100%",
+                        ),
+                        comps.create_selection(
+                            id=data_extension_id,
+                            label="Choose a file extension for the example data",
+                            choices=[e for e in PropertyExtension],
+                            selected=PropertyExtension.CSV,
+                            width="100%",
+                        ),
+                    ),
+                ),
+                ui.column(
+                    8,
+                    ui.output_ui(data_card_id),
+                ),
+            ),
+            size="l",
+            easy_close=True,
+            footer=ui.modal_button(
+                "Close",
+                class_="btn btn-primary",
+            ),
+        )
+    )
 
 
 def phase_input(n: int) -> ui.Tag:
@@ -47,546 +118,369 @@ def phase_input(n: int) -> ui.Tag:
     )
 
 
-@module.ui
-def ui_() -> ui.Tag:
-    sidebar = comps.create_sidebar(
-        ui.tags.div(
-            views.create_help("help"),
-            comps.group_ui_elements(
-                comps.create_selection(
-                    id="dim",
-                    label="Choose a dimension",
-                    choices=[d for d in Dimension],
-                    selected=Dimension.THREE_D,
-                ),
-                ui.output_ui("box_dim"),
-                ui.output_ui("periodicity"),
-                title="Box dimension",
-                help_text=views.box_help_text(),
-            ),
-            comps.group_ui_elements(
-                comps.create_selection(
-                    id="phase",
-                    label="Choose a phase",
-                    choices=[p for p in Phase],
-                    selected=Phase.SINGLE,
-                ),
-                ui.panel_conditional(
-                    f"input.phase === '{Phase.SINGLE}'",
-                    ui.tags.div(
-                        ui.row(
-                            ui.column(
-                                6,
-                                ui.input_numeric(
-                                    id="single_phase_n_grains",
-                                    label="Number of grains",
-                                    value=1000,
-                                ),
-                            ),
-                            ui.column(
-                                6, views.create_dist_selection(id="single_phase_dist")
-                            ),
-                        ),
-                        ui.output_ui("single_phase_dist_param"),
-                    ),
-                ),
-                ui.panel_conditional(
-                    f"input.phase === '{Phase.UPLOAD}'",
-                    ui.tags.div(
-                        comps.create_upload_handler(
-                            "uploaded_volumes",
-                            "Upload volumes as a csv or txt file",
-                        ),
-                        ui.output_ui("uploaded_volumes_summary"),
-                    ),
-                ),
-                ui.panel_conditional(
-                    f"input.phase === '{Phase.DUAL}'",
-                    ui.tags.div(
-                        phase_input(n=1),
-                        ui.output_ui("phase1_dist_param"),
-                        ui.hr(),
-                        phase_input(n=2),
-                        ui.output_ui("phase2_dist_param"),
-                    ),
-                ),
-                ui.output_text("d90_text"),
-                ui.output_text("volume_percentage_text"),
-                title="Grains",
-                help_text=views.grains_help_text(),
-            ),
-            comps.group_ui_elements(
-                comps.create_selection(
-                    id="seeds_init",
-                    label="Choose how seeds are initialized",
-                    choices=[i for i in SeedInitializer],
-                    selected=SeedInitializer.RANDOM,
-                ),
-                ui.panel_conditional(
-                    f"input.seeds_init === '{SeedInitializer.RANDOM}'",
-                    ui.tags.div(
-                        ui.input_numeric(
-                            id="random_state",
-                            label="Seeds random state",
-                            value=None,
-                        ),
-                        ui.help_text(
-                            "Seeds will be randomly generated in the specified box above."
-                        ),
-                    ),
-                ),
-                ui.panel_conditional(
-                    f"input.seeds_init === '{SeedInitializer.UPLOAD}'",
-                    ui.tags.div(
-                        comps.create_upload_handler(
-                            "uploaded_seeds",
-                            "Uplaod seeds as a csv or txt file",
-                        ),
-                        ui.output_ui("uploaded_seeds_summary"),
-                    ),
-                ),
-                views.create_numeric_input(
-                    ["tol", "n_iter", "damp_param"],
-                    ["Volume tolerance", "Lloyd iterations", "Damp param"],
-                    [1.0, 5, 1.0],
-                ),
-                title="Algorithm",
-                help_text=views.algo_help_text(),
-            ),
-            comps.create_input_task_button(
-                id="generate",
-                label="Generate microstructure",
-                icon="person-running",
-            ),
-            ui.input_dark_mode(mode="light"),
-            views.feedback_text(),
-            views.app_version_text(),
+def sidebar() -> ui.Tag:
+    return comps.group_ui_elements(
+        comps.create_selection(
+            id="phase",
+            label="Choose a phase",
+            choices=[p for p in Phase],
+            selected=Phase.SINGLE,
         ),
-        id="sidebar",
+        ui.panel_conditional(
+            f"input.phase === '{Phase.SINGLE}'",
+            ui.tags.div(
+                ui.row(
+                    ui.column(
+                        6,
+                        ui.input_numeric(
+                            id="single_phase_n_grains",
+                            label="Number of grains",
+                            value=1000,
+                        ),
+                    ),
+                    ui.column(6, views.create_dist_selection(id="single_phase_dist")),
+                ),
+                ui.output_ui("single_phase_dist_param"),
+            ),
+        ),
+        ui.panel_conditional(
+            f"input.phase === '{Phase.UPLOAD}'",
+            ui.tags.div(
+                comps.create_upload_handler(
+                    "uploaded_volumes",
+                    "Upload volumes as a csv or txt file",
+                ),
+                ui.output_ui("uploaded_volumes_summary"),
+            ),
+        ),
+        ui.panel_conditional(
+            f"input.phase === '{Phase.DUAL}'",
+            ui.tags.div(
+                phase_input(n=1),
+                ui.output_ui("phase1_dist_param"),
+                ui.hr(),
+                phase_input(n=2),
+                ui.output_ui("phase2_dist_param"),
+            ),
+        ),
+        ui.output_text("d90_text"),
+        ui.output_text("volume_percentage_text"),
+        title="Grains",
+        help_text=views.grains_help_text(),
     )
 
-    return comps.create_page_sidebar(common.page_ui("common"), sidebar=sidebar)
+
+def add_dist_params(iv: InputValidator, dist: str, id_prefix: str, **kwargs) -> dict:
+    match dist:
+        case Distribution.UNIFORM:
+            iv.add_rule(f"{id_prefix}_low", utils.req_gt(rhs=0))
+            iv.add_rule(f"{id_prefix}_high", utils.req_gt(rhs=0))
+
+            return {
+                k: v() for k, v in zip(("low", "high"), kwargs[Distribution.UNIFORM])
+            }
+
+        case Distribution.LOGNORMAL:
+            iv.add_rule(f"{id_prefix}_mean", utils.req_gt(rhs=0))
+            iv.add_rule(f"{id_prefix}_std", utils.req_gt(rhs=0))
+
+            return {
+                k: v() for k, v in zip(("mean", "std"), kwargs[Distribution.LOGNORMAL])
+            }
+
+        case _:
+            return {}
 
 
-@module.server
-def server(input: Inputs, output: Outputs, session: Session) -> None:
-    iv = InputValidator()
-    iv.add_rule("length", utils.req_gt(rhs=0))
-    iv.add_rule("breadth", utils.req_gt(rhs=0))
+def create_dist_params(
+    dist: str,
+    id_prefix: str,
+) -> ui.Tag:
+    text = f"{dist} distribution selected;"
+    match dist:
+        case Distribution.CONSTANT:
+            return ui.help_text(f"{text} all volumes will be equal for this phase.")
+
+        case Distribution.UNIFORM:
+            return ui.tags.div(
+                views.create_numeric_input(
+                    ids=[f"{id_prefix}_{p}" for p in ("low", "high")],
+                    labels=["Low", "High"],
+                    defaults=[1, 2],
+                ),
+                ui.help_text(
+                    f"{text} volumes will be distibuted uniformly in [Low, High)."
+                ),
+            )
+
+        case Distribution.LOGNORMAL:
+            return ui.tags.div(
+                views.create_numeric_input(
+                    ids=[f"{id_prefix}_{p}" for p in ("mean", "std")],
+                    labels=["ECD Mean", "ECD Std"],
+                    defaults=[1, 0.35],
+                ),
+                ui.help_text(
+                    f"""{text} Equivalent Circle Diameters (ECDs) will be sampled from
+                    a lorgnormal distribution with mean 'ECD mean' and standard deviation 'ECD std'.
+                    """
+                ),
+            )
+
+        case _:
+            raise ValueError(
+                f"Mismatch dist: {dist}. Input must be one of {', '.join(Distribution)}."
+            )
+
+
+def volumes(
+    input: Inputs,
+    iv: InputValidator,
+    domain_vol: float,
+    space_dim: int,
+    uploaded_volumes: pd.DataFrame | None,
+) -> np.ndarray | Exception:
     iv.add_rule("tol", utils.req_gt(rhs=0))
-    iv.add_rule("damp_param", utils.req_between(left=0.0, right=1.0))
-    iv.add_rule("n_iter", utils.req_int_gte(rhs=0))
 
-    def add_dist_param_to_iv(dist: str, id_prefix: str, **kwargs) -> dict:
-        match dist:
-            case Distribution.UNIFORM:
-                iv.add_rule(f"{id_prefix}_low", utils.req_gt(rhs=0))
-                iv.add_rule(f"{id_prefix}_high", utils.req_gt(rhs=0))
+    match input.phase():
+        case Phase.SINGLE:
+            iv.add_rule("single_phase_n_grains", utils.req_int_gt(rhs=0))
 
-                return {
-                    k: v()
-                    for k, v in zip(("low", "high"), kwargs[Distribution.UNIFORM])
-                }
-
-            case Distribution.LOGNORMAL:
-                iv.add_rule(f"{id_prefix}_mean", utils.req_gt(rhs=0))
-                iv.add_rule(f"{id_prefix}_std", utils.req_gt(rhs=0))
-
-                return {
-                    k: v()
-                    for k, v in zip(("mean", "std"), kwargs[Distribution.LOGNORMAL])
-                }
-
-            case _:
-                return {}
-
-    # define reactive variables for holding uploaded seeds and volumes
-    _uploaded_seeds = reactive.Value(value=None)
-    _uploaded_volumes = reactive.Value(value=None)
-
-    @reactive.effect
-    def _() -> None:
-        file: list[FileInfo] | None = input.uploaded_seeds()
-        if file is not None:
-            _uploaded_seeds.set(pd.read_csv(file[0]["datapath"]))
-
-    @reactive.effect
-    def _() -> None:
-        file: list[FileInfo] | None = input.uploaded_volumes()
-
-        if file is not None:
-            _uploaded_volumes.set(pd.read_csv(file[0]["datapath"]))
-
-    @reactive.calc
-    @reactive.event(input.generate)
-    def _fitted_data() -> tuple[SynthetMicData, LaguerreDiagramGenerator] | Exception:
-        state_vars = {}
-
-        box_dim = utils.parse_box_dim(
-            utils.Box(
-                length=input.length(), breadth=input.breadth(), height=input.height()
-            ),
-            dim=input.dim(),
-        )
-        periodic = utils.parse_periodicity(
-            utils.Periodic(
-                is_x_periodic=input.is_x_periodic(),
-                is_y_periodic=input.is_y_periodic(),
-                is_z_periodic=input.is_z_periodic(),
-            ),
-            dim=input.dim(),
-        )
-
-        if input.dim() == Dimension.THREE_D:
-            iv.add_rule("height", utils.req_gt(rhs=0))
-
-        state_vars["periodic"] = periodic
-        state_vars["space_dim"] = len(box_dim)
-        state_vars["box_dim"] = box_dim
-        state_vars["domain_vol"] = np.prod(box_dim)
-        state_vars["domain"] = np.array([[0, d] for d in box_dim])
-
-        match input.phase():
-            case Phase.SINGLE:
-                iv.add_rule("single_phase_n_grains", utils.req_int_gt(rhs=0))
-
-                kwargs = add_dist_param_to_iv(
-                    input.single_phase_dist(),
-                    "single_phase",
-                    **dict(
-                        zip(
-                            [Distribution.UNIFORM, Distribution.LOGNORMAL],
-                            [
-                                (input.single_phase_low, input.single_phase_high),
-                                (input.single_phase_mean, input.single_phase_std),
-                            ],
-                        )
-                    ),
-                )
-
-                iv.enable()
-                if iv.is_valid():
-                    volumes = utils.sample_single_phase_vols(
-                        input.single_phase_dist(),
-                        input.single_phase_n_grains(),
-                        state_vars["domain_vol"],
-                        space_dim=state_vars["space_dim"],
-                        **kwargs,
+            kwargs = add_dist_params(
+                iv=iv,
+                dist=input.single_phase_dist(),
+                id_prefix="single_phase",
+                **dict(
+                    zip(
+                        [Distribution.UNIFORM, Distribution.LOGNORMAL],
+                        [
+                            (input.single_phase_low, input.single_phase_high),
+                            (input.single_phase_mean, input.single_phase_std),
+                        ],
                     )
-
-                    state_vars["n_grains"] = input.single_phase_n_grains()
-                    state_vars["volumes"] = volumes
-
-            case Phase.DUAL:
-                for n in (1, 2):
-                    iv.add_rule(f"phase{n}_n_grains", utils.req_int_gt(rhs=0))
-                    iv.add_rule(f"phase{n}_vol_ratio", utils.req_gt(rhs=0))
-
-                phase1_kwargs = add_dist_param_to_iv(
-                    input.phase1_dist(),
-                    "phase1",
-                    **dict(
-                        zip(
-                            [Distribution.UNIFORM, Distribution.LOGNORMAL],
-                            [
-                                (input.phase1_low, input.phase1_high),
-                                (input.phase1_mean, input.phase1_std),
-                            ],
-                        )
-                    ),
-                )
-                phase2_kwargs = add_dist_param_to_iv(
-                    input.phase2_dist(),
-                    "phase2",
-                    **dict(
-                        zip(
-                            [Distribution.UNIFORM, Distribution.LOGNORMAL],
-                            [
-                                (input.phase2_low, input.phase2_high),
-                                (input.phase2_mean, input.phase2_std),
-                            ],
-                        )
-                    ),
-                )
-
-                iv.enable()
-                if iv.is_valid():
-                    volumes = utils.sample_dual_phase_vols(
-                        (input.phase1_dist(), input.phase2_dist()),
-                        (input.phase1_n_grains(), input.phase2_n_grains()),
-                        (input.phase1_vol_ratio(), input.phase2_vol_ratio()),
-                        state_vars["domain_vol"],
-                        state_vars["space_dim"],
-                        (phase1_kwargs, phase2_kwargs),
-                    )
-
-                    state_vars["n_grains"] = sum(
-                        [input.phase1_n_grains(), input.phase2_n_grains()]
-                    )
-                    state_vars["volumes"] = volumes
-
-            case Phase.UPLOAD:
-                iv.enable()
-                if iv.is_valid():
-                    if _uploaded_volumes() is None:
-                        return Exception(
-                            "Grain volumes not uploaded. Upload grain volumes and try again."
-                        )
-
-                    # validate the volumes
-                    val_out = utils.validate_df(
-                        _uploaded_volumes(),
-                        expected_colnames=[utils.VOLUMES],
-                        expected_dim=None,
-                        expected_type="float",
-                        file="volumes",
-                        bounds=None,
-                    )
-                    if isinstance(val_out, str):
-                        return Exception(val_out)
-
-                    # check if domain volume is close to the sum of the uploaded volumes.
-                    VOL_DIFF_TOL = 1e-6
-                    volumes = _uploaded_volumes()[
-                        utils.VOLUMES
-                    ].values  # get the underlying numpy array
-                    diff = abs(volumes.sum() - state_vars["domain_vol"])
-                    if diff > VOL_DIFF_TOL:
-                        return Exception(f"""Mismatch total volume: domain volume is
-                        {state_vars["domain_vol"]} whereas total uploaded volume is {volumes.sum()};
-                        a difference of {diff}. Volume difference must be at most {VOL_DIFF_TOL:.2e}.
-                        """)
-
-                    state_vars["n_grains"] = len(volumes)
-                    state_vars["volumes"] = volumes
-
-            case _:
-                return Exception(
-                    f"Mismatch phase: {input.phase()}. Input must be one of {', '.join(Phase)}."
-                )
-
-        # check the validity of user inputs
-        if not {"n_grains", "volumes"}.issubset(state_vars):
-            return Exception(
-                "Invalid inputs. Please check all fields for the required values."
+                ),
             )
 
-        # deal with the seeds
-        match input.seeds_init():
-            case SeedInitializer.RANDOM:
-                iv.add_rule("random_state", utils.int_gte(rhs=0, allow_none=True))
-                iv.enable()
+            iv.enable()
+            if not iv.is_valid():
+                return Exception(views.invalid_input_text())
 
-                if not iv.is_valid():
-                    return Exception(
-                        "Invalid random seed. Please check your input and try again."
-                    )
-
-                seeds = sample_random_seeds(
-                    domain=state_vars["domain"],
-                    n_grains=state_vars["n_grains"],
-                    random_state=input.random_state(),
-                )
-                state_vars["seeds"] = seeds
-
-            case SeedInitializer.UPLOAD:
-                if _uploaded_seeds() is None:
-                    return Exception("Seeds not uploaded. Upload seeds and try again.")
-
-                # validate the seeds
-                val_out = utils.validate_df(
-                    _uploaded_seeds(),
-                    expected_colnames=list(utils.COORDINATES)[
-                        : state_vars["space_dim"]
-                    ],
-                    expected_dim=(
-                        state_vars["n_grains"],
-                        state_vars["space_dim"],
-                    ),
-                    expected_type="float",
-                    file="seeds",
-                    bounds=dict(
-                        zip(
-                            utils.COORDINATES[: state_vars["space_dim"]],
-                            state_vars["domain"],
-                        )
-                    ),
-                )
-                if isinstance(val_out, str):
-                    return Exception(val_out)
-
-                # add domain and seeds to state_vars since they have been uploaded
-                seeds = _uploaded_seeds()
-                state_vars["seeds"] = seeds.values
-
-            case _:
-                return Exception(
-                    f"Mismatch seed initializer: {input.seeds_init()}. Input must be one of {', '.join(SeedInitializer)}."
-                )
-
-        # finally check if volumes and seeds dim match; very important for the uploads
-        if len(state_vars["volumes"]) != len(state_vars["seeds"]):
-            return Exception(
-                f"""The number of samples in seeds and grain volumes do not match:
-              len(seeds)={len(state_vars["seeds"])}, len(volumes)={len(state_vars["volumes"])}"""
+            volumes = utils.sample_single_phase_vols(
+                dist=input.single_phase_dist(),
+                n_grains=input.single_phase_n_grains(),
+                domain_vol=domain_vol,
+                space_dim=space_dim,
+                **kwargs,
             )
 
-        return utils.fit(
-            domain=state_vars["domain"],
-            seeds=state_vars["seeds"],
-            volumes=state_vars["volumes"],
-            periodic=list(state_vars["periodic"]),
-            tol=float(input.tol()),
-            n_iter=input.n_iter(),
-            damp_param=float(input.damp_param()),
-        )
+            return volumes
 
-    @reactive.effect
-    @reactive.event(input.generate)
-    async def _():
-        await session.send_custom_message(
-            "ga_event",
-            asdict(
-                utils.Event(
-                    name="laguerre_generate",
-                    category="button",
-                    label="generate laguerre diagram",
+        case Phase.DUAL:
+            for n in (1, 2):
+                iv.add_rule(f"phase{n}_n_grains", utils.req_int_gt(rhs=0))
+                iv.add_rule(f"phase{n}_vol_ratio", utils.req_gt(rhs=0))
+
+            phase1_kwargs = add_dist_params(
+                iv=iv,
+                dist=input.phase1_dist(),
+                id_prefix="phase1",
+                **dict(
+                    zip(
+                        [Distribution.UNIFORM, Distribution.LOGNORMAL],
+                        [
+                            (input.phase1_low, input.phase1_high),
+                            (input.phase1_mean, input.phase1_std),
+                        ],
+                    )
+                ),
+            )
+            phase2_kwargs = add_dist_params(
+                iv=iv,
+                dist=input.phase2_dist(),
+                id_prefix="phase2",
+                **dict(
+                    zip(
+                        [Distribution.UNIFORM, Distribution.LOGNORMAL],
+                        [
+                            (input.phase2_low, input.phase2_high),
+                            (input.phase2_mean, input.phase2_std),
+                        ],
+                    )
+                ),
+            )
+
+            iv.enable()
+            if not iv.is_valid():
+                return Exception(views.invalid_input_text())
+
+            volumes = utils.sample_dual_phase_vols(
+                dist=(input.phase1_dist(), input.phase2_dist()),
+                n_grains=(input.phase1_n_grains(), input.phase2_n_grains()),
+                vol_ratio=(input.phase1_vol_ratio(), input.phase2_vol_ratio()),
+                domain_vol=domain_vol,
+                space_dim=space_dim,
+                dist_kwargs=(phase1_kwargs, phase2_kwargs),
+            )
+
+            return volumes
+
+        case Phase.UPLOAD:
+            iv.enable()
+            if not iv.is_valid():
+                return Exception(views.invalid_input_text())
+
+            if uploaded_volumes is None:
+                return Exception(
+                    "Grain volumes not uploaded. Upload grain volumes and try again."
                 )
-            ),
-        )
 
-    @render.download(
-        filename=lambda: f"synthetmic-gui-example-{input.example_data()}.zip",
-        media_type="application/zip",
+            val_out = utils.validate_df(
+                uploaded_volumes,
+                expected_colnames=[utils.VOLUMES],
+                expected_dim=None,
+                expected_type="float",
+                file="volumes",
+                bounds=None,
+            )
+            if isinstance(val_out, str):
+                return Exception(val_out)
+
+            VOL_DIFF_TOL = 1e-6
+            volumes = uploaded_volumes[utils.VOLUMES].values
+            diff = abs(volumes.sum() - domain_vol)
+            if diff > VOL_DIFF_TOL:
+                return Exception(f"""Mismatch total volume: domain volume is
+                    {domain_vol} whereas total uploaded volume is {volumes.sum()};
+                    a difference of {diff}. Volume difference must be at most {VOL_DIFF_TOL:.2e}.
+                    """)
+
+            return volumes
+
+        case _:
+            return Exception(
+                f"Mismatch phase: {input.phase()}. Input must be one of {', '.join(Phase)}."
+            )
+
+
+def create_example_data_card(
+    name: str,
+    image_id: str,
+    download_id: str,
+) -> ui.Tag:
+    common_tags = """
+        volume tolerance=1.0%, Lloyd iterations=20, damp param=1.0, 
+        colorby=fitted volumes, colormap=plasma.
+        """
+
+    match name:
+        case ExampleDataName.BASIC:
+            info = """
+                This is an example of a basic synthetic microstructure with a 
+                random seed initialization and constant volume distribution.
+                """
+            tags = "2D, random seed, constant volumes, "
+
+        case ExampleDataName.RANDOM:
+            info = """
+                Random distribution of initial seed locations. Here the initial
+                generator locations of the large and small grains are uniformly
+                distributed over the corresponding domain.
+                """
+            tags = "2D, dual-phase volumes, random seeds, "
+
+        case ExampleDataName.BANDED:
+            info = """
+                Banded distribution of initial seed locations.
+                Here, the different sized grains have initial generator
+                locations that lie inside bands within the domain.
+                The sizes of the bands have been chosen so that there are
+                approximately equal numbers of small grains within each
+                small-grain band and approximately equal numbers of large grains
+                within each large-grain band.
+                """
+            tags = "2D, dual-phase volumes, banded seeds, "
+
+        case ExampleDataName.CLUSTERED:
+            info = """
+                Clustered distribution of initial seed locations. 
+                Here, the smaller grains have initial generator locations
+                that lie inside non-overlapping discs.
+                """
+            tags = "2D, dual-phase volumes, clustered seeds, "
+
+        case ExampleDataName.MIXED:
+            info = """
+                A mixed distribution: the initial generators are
+                such that the larger grains are arranged in bands and
+                the smaller grains are a combination of the banded and random distributions.
+                """
+            tags = "2D, dual-phase volumes, mixed seeds, "
+
+        case ExampleDataName.INCREASING:
+            info = """
+                The initial seed locations are distributed such that the 𝑥-coordinate
+                increases with grain size.
+                """
+            tags = "2D, multi-phase volumes, increasing grain size, "
+
+        case ExampleDataName.MIDDLE:
+            info = """
+                The initial seed locations are distributed such that
+                the larger grains are found in the middle of the domain.
+                """
+            tags = "2D, multi-phase volumes, divergent grain size, "
+
+        case ExampleDataName.DP:
+            info = """
+                An RVE of a dual-phase material with a banded microstructure.
+                """
+            tags = "3D, RVE, dual-phase, banded structure, "
+
+        case ExampleDataName.LOGNORMAL:
+            info = """
+                An RVE in which the grain volumes have approximately lognormal distribution
+                The coefficient of variation of the volumes (the ratio of the standard
+                deviation to the mean) is 1.4.
+                """
+            tags = "3D, RVE, lognormal volumes, banded structure, "
+
+        case ExampleDataName.EBSD:
+            info = """
+            In this example we fit a Laguerre diagram to an EBSD image of a
+            single-phase steel. The 'target volumes'
+            are the areas of the grains in the EBSD image. The 'seeds' are the centroids of the grains
+            in the EBSD image. The EBSD data is taken from this [paper](https://doi.org/10.1051/m2an/2025004).
+            """
+            tags = "2D, non-periodic, volume upload, seed upload, volume tolerance = 1, Lloyd iterations = 0, damp param = 1"
+
+        case _:
+            raise ValueError(
+                f"Invalid data name '{name}'; name must be one of [{', '.join(ExampleDataName)}]."
+            )
+
+    return ui.card(
+        ui.card_header(name),
+        ui.output_image(image_id, fill=True, height="100%", width="100%"),
+        ui.markdown(info),
+        ui.help_text(
+            f"Tags: {tags}{'' if name == ExampleDataName.EBSD else common_tags}"
+        ),
+        comps.create_download_button(
+            id=download_id,
+            label="Download data",
+        ),
     )
-    def download_example_data() -> Generator[bytes, None, None]:
-        yield utils.create_example_data_bytes(
-            name=input.example_data(), file_extension=input.example_data_extension()
-        )
 
-    @render.image
-    def example_data_image() -> ImgData:
-        return utils.load_image(
-            pathlib.Path().resolve()
-            / "assets"
-            / "imgs"
-            / f"{input.example_data()}.png",
-        )
 
-    @render.ui
-    def example_data_card() -> ui.Tag:
-        return views.create_example_data_card(
-            name=input.example_data(),
-            image_id="example_data_image",
-            download_id="download_example_data",
-        )
+def compute_d90_text(mean: float, std: float) -> str | None:
+    text = """
+            D90 for the lognormal distribution of ECDs with mean {}
+            and std {} is {}. 
+            """
+    if any([i is None for i in (mean, std)]) or mean == 0:
+        return
 
-    @render.ui
-    def box_dim() -> ui.Tag:
-        return views.box_dim(input.dim())
+    d90 = utils.qp(mean=mean, std=std, p=0.9)
+    d90 = utils.format_to_standard_form(d90, precision=2)
 
-    @render.ui
-    def periodicity() -> ui.Tag:
-        return views.periodicity(input.dim())
-
-    @render.table
-    def seeds_summary_table() -> pd.DataFrame:
-        return utils.summarize_df(_uploaded_seeds())
-
-    @render.ui
-    def uploaded_seeds_summary() -> ui.Tag:
-        return views.seeds_summary(_uploaded_seeds(), "seeds_summary_table")
-
-    @render.ui
-    def single_phase_dist_param() -> ui.Tag:
-        return views.create_dist_param(input.single_phase_dist(), "single_phase")
-
-    @render.ui
-    def phase1_dist_param() -> ui.Tag:
-        return views.create_dist_param(input.phase1_dist(), "phase1")
-
-    @render.ui
-    def phase2_dist_param() -> ui.Tag:
-        return views.create_dist_param(input.phase2_dist(), "phase2")
-
-    @render.table
-    def volumes_summary_table() -> pd.DataFrame:
-        return utils.summarize_df(_uploaded_volumes())
-
-    @render.ui
-    def uploaded_volumes_summary() -> ui.Tag:
-        if _uploaded_volumes() is None:
-            return ui.help_text(
-                "No volumes uploaded yet. Information about volumes will be displayed here after upload."
-            )
-
-        return ui.output_table("volumes_summary_table")
-
-    @render.text
-    def volume_percentage_text() -> str | None:
-        if any(
-            [
-                input.phase1_vol_ratio() is None,
-                input.phase2_vol_ratio() is None,
-                input.phase() == Phase.UPLOAD,
-            ]
-        ):
-            return
-
-        if input.phase() == Phase.SINGLE:
-            return "Single phase volume percentage is 100% of the domain volume."
-
-        phase1_vol_percent = (
-            input.phase1_vol_ratio()
-            * 100
-            / (input.phase1_vol_ratio() + input.phase2_vol_ratio())
-        )
-
-        return (
-            f"Phase 1 volume percentage is {phase1_vol_percent:.2f}%; Phase 2 volume percentage is {100.0 - phase1_vol_percent:.2f}% "
-            "of the domain volume."
-        )
-
-    @render.text
-    def d90_text() -> str | None:
-        if (
-            input.phase() == Phase.SINGLE
-            and input.single_phase_dist() == Distribution.LOGNORMAL
-        ):
-            return views.compute_d90_text(
-                mean=input.single_phase_mean(), std=input.single_phase_std()
-            )
-
-        if input.phase() == Phase.DUAL:
-            msg = []
-            for i, p in enumerate((input.phase1_dist(), input.phase2_dist()), start=1):
-                if p == Distribution.LOGNORMAL:
-                    text = views.compute_d90_text(
-                        mean=getattr(input, f"phase{i}_mean")(),
-                        std=getattr(input, f"phase{i}_std")(),
-                    )
-                    if text is not None:
-                        msg.append(f"Phase {i} {text}")
-
-            if msg:
-                return " ".join(msg)
-
-    @reactive.effect
-    @reactive.event(input.help)
-    def _() -> None:
-        views.laguerre_help(
-            data_selection_id="example_data",
-            data_extension_id="example_data_extension",
-            data_card_id="example_data_card",
-        )
-
-    @reactive.effect
-    @reactive.event(input.generate)
-    def _():
-        f = _fitted_data()
-        if isinstance(f, Exception):
-            comps.create_error_notification(str(f))
-
-        else:
-            ui.update_sidebar(id="sidebar", show=False)
-            common.server("common", f, input.generate)
+    return text.format(mean, std, d90)
